requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.10.6
pillow==11.0.0
pytesseract==0.3.10
python-multipart==0.0.9
openai==1.35.10
jinja2==3.1.4
httpx==0.27.0
pytest==8.2.2
requests==2.32.3